fix(pcs): keep the open_columns weight gather integer-typed for an empty block (#539)
## Description

`open_columns` (from #512) gathers each column's weight vector with
`weights[:, np.array(opening_pos[:b])]`. When a block is empty — the
wired DEEP opening passes an all-cubic quotient column with **no base
block** (`base_cols` shape `(N, 0)`) — `opening_pos[:b]` is `[]` and
`np.array([])` defaults to `float64`, which frx rejects: `TypeError:
Indexer must have integer or boolean type, got indexer with type
float64`. #512's tests always had both a base and an extension block, so
the empty-block shape was uncovered.

Pin the gather indices to `np.intp`. The regression test adds the
single-field-block cases (all base, all extension) alongside the
existing mixed-block one.

Found while bumping the downstream consumer (fractalyze/zisk-zorch#97):
zisk's `deep/opening_test` opens a lone cubic column and hit this
immediately.

## Related Issues/PRs

Fixes a regression in #512. Unblocks fractalyze/zisk-zorch#97 and #68.

## Checklist

- [x] Branch name follows Branch Guideline
- [x] Commit messages follow Commit Message Guideline
- [x] Checked Pull Request Guideline

+++ zorch/pcs/deep.py
@@ -98,8 +98,8 @@ def open_columns(
     b = base_cols.shape[1]
     base_sub = base_cols[::stride]
     ext_sub = ext_cols[::stride]
-    base_w = weights[:, np.array(opening_pos[:b])]  # (N, B)
-    ext_w = weights[:, np.array(opening_pos[b:])]  # (N, C)
+    base_w = weights[:, np.array(opening_pos[:b], dtype=np.intp)]  # (N, B)
+    ext_w = weights[:, np.array(opening_pos[b:], dtype=np.intp)]  # (N, C)
     base_openings = fnp.sum(base_w * base_sub, axis=0)
     ext_openings = fnp.sum(ext_w * ext_sub, axis=0)
     return fnp.concatenate([base_openings, ext_openings])
